requirements.txt
streamlit
python-pptx
lxml
